mudando uma coisa para byte
porque não?

+++ src/oop/BicicletaInterface.java
@@ -4,8 +4,8 @@ public interface BicicletaInterface {
     int getMarcha();
     void setMarcha(int novaMarcha);
     int mudaMarcha(int novaMarcha);
-    int getCadencia();
-    void setCadencia(int cadencia);
+    byte getCadencia();
+    void setCadencia(byte cadencia);
     int getVelocidade();
     void aumentaAVelocidade(int aumentoDeVelocidade);
     void freia(int perdaDeVelocidade);
